requirements.txt
chromadb==0.4.22
deep_phonemizer==0.0.19
Flask==3.0.1
inflect==6.0.5
keyboard==0.13.5
langchain==0.1.1
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.3
openai==1.8.0
openai_whisper==20230918
pandas==2.0.3
Pillow==10.1.0
Pillow==10.2.0
playsound==1.2.2
PyAudio==0.2.14
pydub==0.25.1
pygame==2.5.2
scipy==1.11.4
sounddevice==0.4.6
SpeechRecognition==3.10.1
streamlit==1.29.0
torch==2.0.1
tqdm==4.65.2
Unidecode==1.3.7
